regularization: add plotting

+++ MainAnalysis/src/regularization.C
@@ -126,7 +126,7 @@ int regularization(char const* config, char const* selections, char const* outpu
     apply_style(p, cms, system_tag);
     p->set(paper::flags::logy);
 
-    for (int64_t i = 0; i < files.size(); ++i) {
+    for (size_t i = 0; i < files.size(); ++i) {
         p->add((*mse)[i], "MSE", algorithm, prior, object, label);
     }
